Fix some warnings in MSVC.

+++ src/fileaccess.cpp
@@ -231,7 +231,7 @@ void FileAccess::setFile(const QFileInfo& fi, FileAccess* pParent)
 #ifndef Q_OS_WIN 
             // Unfortunately Qt5 symLinkTarget/readLink always returns an absolute path, even if the link is relative
             char s[PATH_MAX + 1];
-            int len = readlink(QFile::encodeName(fi.absoluteFilePath()).constData(), s, PATH_MAX);
+            ssize_t len = readlink(QFile::encodeName(fi.absoluteFilePath()).constData(), s, PATH_MAX);
             if(len > 0)
             {
                 s[len] = '\0';
